refactor: 프로젝트 구조 변경과 오타 수정
- 프로젝트 구조 변경
